refactor: correction in packages
